Build Instabilities Fix (#554)
* Testing reading the version during tests

* retry running TestWindowsPlan

* Tried to fix an issue in NoAssemblyInfo.TapPlan by installing the current debug version instead of the one downloaded from the build.

added more tests of engine using matrix.

- Moved run parse test to a regression test.
- Added retry with exponential backoff to httprepository.
- Improved exception logging by writing the type of exception.
- Added a performance improvement to threadmanager whenthen  default token is used.
- Added improvement to ProcessStep to working directory can be null or empty.

* build issue fix: Added test

- added test to ensure that result tables can only be merged if their columns has the same parameters.

Moved ResultTableOptimizer so it can be tested.

* Build issue fix: fixed bug

Fixed a bug in result table optimizer that caused builds to fail. This bug occurs when merged tables has different parameters. This has been fixed by checking that the parameters are the same before merging.

* 551 550 Package file installation improvement

- Packages cannot contain / characters. This limitation already exists in the package repository server.
- tap package install with packages ending with .TapPackage are assumed to be package files and hence we dont ask the repositories. This makes it easier to tell the user what went wrong.
- Added comment about '/' to documentation.

* buildissuefix:

- more robust http repository client and less error logs when it is already in error mode.
- Modified PackageSlashTest as we no longer support packages with slashes in the name.

* Disabled repeating regression test plans as we need to handle the issue the first time.

* Added GetClientHandleWithRetryPolicy to avoid conflicting with an already existing APi point.
Fixed issue in regression test plan referencing the Debug folder.

* Fix double pipelines

* - Removed Slash test since we no longer support that use case.
- Fixed API change.

* Reduced the amount of testing. Added more linux testing.

* Build Issue Fix:

- added check when copying to package cache, so that if the file already exists there it is not overwritten. This can otherwise cause issues when multiple tap processes are trying to install at the same time.

- Minor adjustment in ProcessStep to gracefully exit.

* Cleanup:

Removed some unused variables

* Improved possible bug in retry logic in HttpPackageRepository.cs

Improved code structure for retry logic in HttpHandler.

* build issue fix: Minor cleanup

+++ BasicSteps/ProcessStep.cs
@@ -28,7 +28,7 @@ namespace OpenTap.Plugins.BasicSteps
 
         [Display("Working Directory", Order: -2.3, Description: "The directory where the program will be started in.")]
         [DirectoryPath]
-        public string WorkingDirectory { get; set; } = Directory.GetCurrentDirectory();
+        public string WorkingDirectory { get; set; } = "";
 
         [Display("Wait For Process to End", Order: -2.2,
             Description: "Wait for the process to terminate before continuing.")]
@@ -108,10 +108,11 @@ namespace OpenTap.Plugins.BasicSteps
                 {
                     FileName = Application,
                     Arguments = Arguments,
-                    WorkingDirectory = Path.GetFullPath(WorkingDirectory),
+                    WorkingDirectory = string.IsNullOrEmpty(WorkingDirectory) ? Directory.GetCurrentDirectory() : Path.GetFullPath(WorkingDirectory),
                     UseShellExecute = false,
                     RedirectStandardOutput = true,
                     RedirectStandardError = true,
+                    RedirectStandardInput = true,
                     CreateNoWindow = true
                 }
             };
@@ -120,7 +121,18 @@ namespace OpenTap.Plugins.BasicSteps
                 Log.Debug("Ending process '{0}'.", Application);
                 try
                 {  // process.Kill may throw if it has already exited.
-                    process.Kill();
+                    try
+                    {
+                        // signal to the sub process that no more input will arrive.
+                        // For many process this has the same effect as CTRL+C as stdin is closed.
+                        process.StandardInput.Close();
+                    }
+                    catch
+                    {
+                        // this might be ok. It probably means that the input has already been closed.
+                    }
+                    if (!process.WaitForExit(500)) // give some time for the process to close by itself.
+                        process.Kill();
                 }
                 catch(Exception ex)
                 {
